add more changes before release
