+++ packages/frontend/src/components/NewsDrawer/NewsItemList.ts
@@ -11,7 +11,7 @@ const IF_THEN_EXTERNAL_LINK =
 
 export const NEWS_ITEM_LIST: NewsItemProps[] = [
   {
-    date: '2024-10-03',
+    date: '2024-10-04',
     tag: NEW_FEATURE_TAG,
     title: 'Save time building workflows with Templates!',
     details: `Each template comes with pre-selected steps and content set up for you. Use them as starting points or to get inspired by what others are creating. We are continuously working on adding more relevant templates to speed up workflow building. If you need a template, put in a request [here](${URLS.TEMPLATES_FORM_LINK})!`,
